New proof of work now implemented
